move gui2 overrides into common
