+++ buchschloss/gui2/common.py
@@ -1,6 +1,7 @@
 """shared utilities"""
 
 import tkinter as tk
+import tkinter.messagebox as tk_msg
 import typing
 import types
 import contextlib
@@ -8,6 +9,7 @@ import collections
 from functools import partial
 
 from .. import core
+from .. import utils
 from . import main
 
 
@@ -57,3 +59,41 @@ def destroy_all_children(widget: tk.Widget):
     """Destroy all children of the given widget"""
     for child in tuple(widget.children.values()):
         child.destroy()
+
+# NOTE: the following functions aren't/shouldn't used anywhere.
+# The decorator registers them in NSWithLogin.
+
+
+@NSWithLogin.override('Book', 'new')
+def book_new(**kwargs):
+    tk_msg.showinfo(
+        utils.get_name('Book'),
+        utils.get_name('Book::new_id_{}').format(
+            core.Book.new(login_context=main.app.current_login, **kwargs))
+    )
+
+
+@NSWithLogin.override('Book', 'search')
+def book_search(condition):
+    return core.Book.search(
+        (condition, 'and', ('is_active', 'eq', True)),
+        login_context=main.app.current_login,
+    )
+
+
+@NSWithLogin.override('Borrow', 'restitute')
+def borrow_restitute(book):
+    core.Borrow.edit(
+        NSWithLogin(core.Book).view_ns(book)['borrow'],
+        is_back=True,
+        login_context=main.app.current_login,
+    )
+
+
+@NSWithLogin.override('Borrow', 'extend')
+def borrow_extend(book, weeks):
+    core.Borrow.edit(
+        NSWithLogin(core.Book).view_ns(book)['borrow'],
+        weeks=weeks,
+        login_context=main.app.current_login,
+    )
